Refactor button and heading components; update imports and styles
