--- apps/cms/schemas/arrays/index.ts
@@ -1,1 +1,0 @@
-export { blockContent } from './block-content';
