--- apps/web/app/components/design-system/button/index.ts
@@ -1,3 +1,0 @@
-export { Button, type ButtonProps } from './button';
-export { ButtonLink, type ButtonLinkProps } from './button-link';
-export type { ButtonVariantProps } from './get-button-styles';
