--- apps/web/app/components/design-system/field/index.ts
@@ -1,1 +1,0 @@
-export { FieldMessage, type FieldMessageProps, type Tone } from './field';
