--- apps/web/app/components/design-system/heading/index.ts
@@ -1,4 +1,0 @@
-export type { HeadingVariantProps } from './get-heading-styles';
-export { getHeadingStyles } from './get-heading-styles';
-export type { HeadingProps } from './heading';
-export { Heading } from './heading';
