+++ apps/web/app/components/not-found.tsx
@@ -1,4 +1,4 @@
-import { ButtonLink } from './design-system/button';
+import { ButtonLink } from './design-system/button-link';
 import { Heading } from './design-system/heading';
 import { SplitBackground } from './split-background';
