+++ apps/web/app/routes/_index.tsx
@@ -7,9 +7,9 @@ import { data, useLoaderData } from 'react-router';
 import { z } from 'zod';
 import { BrandsWeLove } from '../components/brands-we-love';
 import { ContactForm } from '../components/contact-form/form';
-import { ButtonLink } from '../components/design-system/button';
+import { ButtonLink } from '../components/design-system/button-link';
+import { getHeadingStyles } from '../components/design-system/get-heading-styles';
 import { Heading } from '../components/design-system/heading';
-import { getHeadingStyles } from '../components/design-system/heading/get-heading-styles';
 import { Divider } from '../components/divider';
 import { StoreLocationMap } from '../components/map';
 import { NewsletterSignup } from '../components/newsletter/form';
