+++ apps/web/app/routes/blog._index.tsx
@@ -3,8 +3,8 @@ import type { LoaderFunctionArgs, MetaFunction } from 'react-router';
 import { data as json, Link, useLoaderData, useLocation } from 'react-router';
 import invariant from 'tiny-invariant';
 import { z } from 'zod';
-import { ButtonLink } from '../components/design-system/button';
-import { getHeadingStyles } from '../components/design-system/heading';
+import { ButtonLink } from '../components/design-system/button-link';
+import { getHeadingStyles } from '../components/design-system/get-heading-styles';
 import { Hero } from '../components/hero';
 import { CACHE_SHORT, routeHeaders } from '../lib/cache';
 import { getBlogPostCount, getBlogPosts } from '../lib/get-blog-posts';
